Refactor workspace config
